requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
prometheus_client==0.24.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
redis==7.3.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
